+++ bot.py
@@ -12,16 +12,7 @@ import json
 
 #Import own classes
 from event import Event
-
-#This is how you Get your Complete TimeStamp
-def gcts():
-    return datetime.now().strftime("%Y.%m.%d %H:%M:%S")
-
-#This is how you log it
-def log(inputstr):
-    print('[' + gcts() + '] ' + inputstr)
-    with open('logs/' + datetime.now().strftime('%Y-%m-%d') + '_moevius.log', 'a') as f:
-        f.write('[' + gcts() + '] ' + inputstr + '\n')
+from myfunc import log
 
 #Loading .env, important for the token
 load_dotenv()
@@ -36,7 +27,10 @@ timenow = ''
 channels = {}
 
 #Create events
-stream = Event('stream')
+events = {
+    'stream': Event('stream'),
+    'game': Event('game')
+}
 
 #Default Settings
 
@@ -48,31 +42,35 @@ def load_file(name):
         pass
 
 def import_files():
-    global settings, responses, server
+    global settings, responses, channels, server
 
     settings = load_file('settings')
     responses = load_file('responses')
 
-    stream.load()
-    log(f'Stream-Zeit wurden geladen: {stream.time}')
-
     #Setup Discord objects after settings are loaded
     #Get guild
+    log(f"Server-Suche startet.")
     server = client.get_guild(int(settings['server_id']))
-    log(f"Server {server.name} gefunden, ID: {settings['server_id']}.")
+    log(f"Server-Suche abgeschlossen: {server.name} - {server.id}.")
 
-    #Get channel for streams
+    #Get channel for stream and games
+    log(f"Channel-Suche startet.")
     for c in server.text_channels:
         if c.name == settings['channels']['stream']:
             channels['stream'] = c
-            log(f"Für den Stream wurde der Channel mit dem Namen {settings['channels']['stream']} gefunden, ID: {c.id}")
-    
-    #Settings loaded
-    log("Die Einstellungen wurden komplett geladen.")
+            log(f"Channel für Stream gefunden: {c.name} - {c.id}")
+        elif c.category != None:
+            if c.category.name == 'Spiele':
+                channels[c.name] = c
+                log(f"Channel für Spiel gefunden: {c.name} - {c.id}")
+    log(f"Channel-Suche abgeschlossen.")
 
 #Importing setting files
 def startup():
-    global settings, responses, server, stream, fragen
+    global fragen, events
+
+    for e in events.values():
+        e.load()
 
     import_files()
     
@@ -83,31 +81,40 @@ def startup():
 
 async def loop():
     #I like global variables a lot
-    global timenow, stream
+    global timenow, events
     
     #Wait until ready
     await client.wait_until_ready()
-    log('Der Loop wurde gestartet.')
+    log("Der Loop wurde gestartet.")
     
     #Endless loop for checking timenow
     while True:
         #Update timenow only if it needs to be updated
         if timenow != datetime.now().strftime('%H:%M'):
             timenow = datetime.now().strftime('%H:%M')
-            #Check for stream now?
-            if stream.time == timenow:
-                log('Der Stream geht los.')
-                await channels['stream'].send('Oh, ist es denn schon ' + stream.time + ' Uhr? Dann ab auf https://www.twitch.tv/schnenko/ ... der Stream fängt an, Krah Krah!')
-                stream.reset()
-                stream.save()
-                log('Post wurde abgesetzt, Timer wurde resettet.')
+
+            #Check for events now
+            for e in events.values():
+                if e.time == timenow:
+                    log(f"Ein Event beginnt: {e.name}!")
+                    if e.name == 'stream':
+                        await channels['stream'].send(f"Oh, ist es denn schon {e.time} Uhr? Dann ab auf https://www.twitch.tv/schnenko/ ... der Stream fängt an, Krah Krah!")
+                    elif e.name == 'game':
+                        await channels['game'].send(f"Oh, ist es denn schon {e.time} Uhr? Dann ab in den Voice-Chat, {e.game} fängt an, Krah Krah!")
+                    e.reset()
+                    e.save()
+                    log('Event-Post abgesetzt, Timer resettet.')
         await asyncio.sleep(10)
 
 @client.event
 async def on_ready():
-    #load settings
+    #Load Settings for the first time
     startup()
     log("Ready to Rumble!")
+
+    #Start the Loop
+    client.loop.create_task(loop())
+
 
 @client.event
 async def on_message(message):
@@ -117,19 +124,26 @@ async def on_message(message):
     
     #Commands
     if message.content.startswith('!'):
+        command = str(message.content.split(' ')[0][1:])
         #Commands for super-users
         if message.author.name in settings['super-users']:
             #Stream-Command
-            if message.content.startswith('stream', 1):
-                time = re.findall(r'\d{2}:\d{2}$|$', message.content)[0]
+            if command in events.keys():
+                time = str(message.content.split(' ')[1])
+
                 if time == '':
-                    await message.channel.send(f"Danke, {message.author.display_name}, ich werde den Reminder zurücksetzen, Krah Krah!")
-                    log(f"{message.author.name} hat den Stream-Zeitpunkt resettet!")
+                    events[command].reset()
+                    await message.channel.send(f"Danke, {message.author.display_name}, ich habe den Reminder zurückgesetzt, Krah Krah!")
+                    log(f"Event resettet: {message.author.name} - {command}")
                 else:
-                    await message.channel.send(f"Danke, {message.author.display_name}, ich werde einen Reminder für {time} Uhr einrichten, Krah Krah!")
-                    log(f"{message.author.name} hat den Stream-Zeitpunkt auf {time} festgelegt und das Savefile wurde geupdatet.!")
-                stream.update(time)
-                stream.save()
+                    if command == 'stream':
+                        events[command].updateStream(time)
+                        await message.channel.send(f"Danke, {message.author.display_name}, ich habe einen Stream-Reminder für {time} Uhr eingerichtet, Krah Krah!")
+                    elif command == 'game':
+                        channels['game'] = message.channel
+                        game = channels['game'].name.replace('-',' ').title()
+                        events[command].updateGame(time, game)
+                        await message.channel.send(f"Danke, {message.author.display_name}, ich habe einen Coop-Reminder für {game} um {time} Uhr eingerichtet, Krah Krah!")
             
             #Reload-Settings
             elif message.content.startswith('reload', 1):
@@ -146,11 +160,11 @@ async def on_message(message):
     #Requests
     elif message.content.startswith('?'):
         if message.content.startswith('wann', 1):
-            if stream.time == '':
+            if events['stream'].time == '':
                 await message.channel.send("Uff, der faule Schnenko hat anscheinend keinen Stream angekündigt, Krah Krah!")
                 log(f"{message.author.name} hat nach dem Stream gefragt, aber es gibt keinen.")
             else:
-                await message.channel.send(f"Gut, dass du fragst! Der nächste Stream beginnt um {stream.time} Uhr, Krah Krah!")
+                await message.channel.send(f"Gut, dass du fragst! Der nächste Stream beginnt um {events['stream'].time} Uhr, Krah Krah!")
                 log(f"{message.author.name} hat nach dem Stream gefragt.")
         #Other requests from responses-file
         elif message.content[1:] in responses['req'].keys():
@@ -168,7 +182,5 @@ async def on_message(message):
                     await message.channel.send(r.format(**locals(), **globals()))
                 log(response['log'].format(**locals(), **globals()))
 
-#Start the Loop
-client.loop.create_task(loop())
 #Connect to Discord
 client.run(TOKEN)